[patch] Adding Prettier

+++ src/classes/Request.ts
@@ -4,7 +4,6 @@ import List from './List'
 import Store from './Store'
 import { GuildGonfig } from './GuildConfig'
 
-
 export default class Request {
 	constructor(
 		list: List,
@@ -29,45 +28,58 @@ export default class Request {
 	readonly prefix: string
 	readonly text: string
 
-	get guild() { return this.message.guild }
-	get channel() { return this.message.channel }
-	get member() { return this.message.member }
-	get author() { return this.message.author }
-	get client() { return this.message.client }
-	get bot() { return this.client.user }
-	get voiceConnection() { return this.guild.voiceConnection }
+	get guild() {
+		return this.message.guild
+	}
+	get channel() {
+		return this.message.channel
+	}
+	get member() {
+		return this.message.member
+	}
+	get author() {
+		return this.message.author
+	}
+	get client() {
+		return this.message.client
+	}
+	get bot() {
+		return this.client.user
+	}
+	get voiceConnection() {
+		return this.guild.voiceConnection
+	}
 
 	get screenname() {
 		return this.message.member.nickname || this.message.author.username
 	}
 
-	async send( text, options?: any ) {
-		const message = await this.message.channel.send( text, options )
-		return Array.isArray( message ) ? message[ 0 ] : message
+	async send(text, options?: any) {
+		const message = await this.message.channel.send(text, options)
+		return Array.isArray(message) ? message[0] : message
 	}
 
-	async sendCode( lang: string, content: any, options?: any ) {
-		return await this.message.channel.sendCode( lang, content )
+	async sendCode(lang: string, content: any, options?: any) {
+		return await this.message.channel.sendCode(lang, content)
 	}
 
-	async reply( text, options?: any ) {
-		return await this.message.reply( text, options )
+	async reply(text, options?: any) {
+		return await this.message.reply(text, options)
 	}
 
-	async destructingReply( text: string ) {
-		return await destructingReply( this.message, text )
+	async destructingReply(text: string) {
+		return await destructingReply(this.message, text)
 	}
 
 	async delete() {
-		if ( this.message.deletable )
-			return await this.message.delete()
+		if (this.message.deletable) return await this.message.delete()
 	}
 
-	async somethingWentWrong( err ) {
-		return somethingWentWrong( this.message, err )
+	async somethingWentWrong(err) {
+		return somethingWentWrong(this.message, err)
 	}
 
 	async missingArguments() {
-		return this.send( 'Please specify an argument for channel' )
+		return this.send('Please specify an argument for channel')
 	}
 }
